+++ Packages/com.unity.ide.rider/Rider/Editor/RiderScriptEditor.cs
@@ -309,12 +309,6 @@ namespace Packages.Rider.Editor
         return false;
       }
       
-      if (path == "" && SystemInfo.operatingSystemFamily == OperatingSystemFamily.MacOSX)
-      {
-        // there is a bug in DllImplementation - use package implementation here instead https://github.cds.internal.unity3d.com/unity/com.unity.ide.rider/issues/21
-        return OpenOSXApp(path, line, column);
-      }
-
       if (!IsUnityScript(path))
       {
         m_ProjectGeneration.SyncIfNeeded(affectedFiles: new string[] { }, new string[] { });
@@ -425,8 +419,11 @@ namespace Packages.Rider.Editor
     {
       if (IsAssetImportWorkerProcess())
         return false;
-      
-#if UNITY_2020_2_OR_NEWER
+
+#if UNITY_2021_1_OR_NEWER
+      if (UnityEditor.MPE.ProcessService.level == UnityEditor.MPE.ProcessLevel.Secondary)
+        return false;
+#elif UNITY_2020_2_OR_NEWER
       if (UnityEditor.MPE.ProcessService.level == UnityEditor.MPE.ProcessLevel.Slave)
         return false;
 #elif UNITY_2020_1_OR_NEWER